initial move of CloseButton to Beta::CloseButton
