+++ lib/primer/view_components/linters/argument_mappers/close_button.rb
@@ -10,12 +10,12 @@ module ERBLint
         ATTRIBUTES = %w[type].freeze
 
         TYPE_OPTIONS = Primer::ViewComponents::Constants.get(
-          component: "Primer::CloseButton",
+          component: "Primer::Beta::CloseButton",
           constant: "TYPE_OPTIONS"
         ).freeze
 
         DEFAULT_TYPE = Primer::ViewComponents::Constants.get(
-          component: "Primer::CloseButton",
+          component: "Primer::Beta::CloseButton",
           constant: "DEFAULT_TYPE"
         ).freeze
