+++ lib/primer/view_components/linters/close_button_component_migration_counter.rb
@@ -14,9 +14,9 @@ module ERBLint
 
       TAGS = %w[button].freeze
       CLASSES = %w[close-button].freeze
-      MESSAGE = "We are migrating close-button to use [Primer::CloseButton](https://primer.style/view-components/components/closebutton), please try to use that instead of raw HTML."
+      MESSAGE = "We are migrating close-button to use [Primer::Beta::CloseButton](https://primer.style/view-components/components/closebutton), please try to use that instead of raw HTML."
       ARGUMENT_MAPPER = ArgumentMappers::CloseButton
-      COMPONENT = "Primer::CloseButton"
+      COMPONENT = "Primer::Beta::CloseButton"
 
       ALLOWED_OCTICON_ARGS = %w[icon aria-label aria].freeze
