+++ static/constants.json
@@ -298,6 +298,13 @@
   },
   "Primer::Beta::Breadcrumbs::Item": {
   },
+  "Primer::Beta::CloseButton": {
+    "DEFAULT_TYPE": "button",
+    "TYPE_OPTIONS": [
+      "button",
+      "submit"
+    ]
+  },
   "Primer::Beta::Flash": {
     "DEFAULT_SCHEME": "default",
     "SCHEME_MAPPINGS": {
@@ -355,11 +362,6 @@
   "Primer::ClipboardCopy": {
   },
   "Primer::CloseButton": {
-    "DEFAULT_TYPE": "button",
-    "TYPE_OPTIONS": [
-      "button",
-      "submit"
-    ]
   },
   "Primer::ConditionalWrapper": {
   },
