--- stories/primer/close_button_stories.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class Primer::CloseButtonStories < ViewComponent::Storybook::Stories
-  layout "storybook_preview"
-
-  story(:close_button) do
-    controls do
-      select(:type, Primer::CloseButton::TYPE_OPTIONS, Primer::CloseButton::DEFAULT_TYPE)
-    end
-  end
-end
